requirements.txt
PyPDF2~=3.0.1
langchain~=0.2.16
huggingface_hub
langchain_community
langchain-huggingface
langchain_redis
langchain-google-genai
faiss-cpu
python-dotenv